made items page
I added the buttons for switching between khareedar and dost modes. I also made navigation for this.
This module has been tested on the front end and needs backend integration. This can be tested by changing the landing page to take you to the items page as soon as the time out occurs.

The bottom buttons have been made a component now.